phase:12 issue #33: refactor random-walk peer-influence architecture modules
